rename public mono/mini subdirectory to mono/jit
To match how embedders see the tree.

Update the runtime to include <mono/jit/jit.h> instead of <mono/mini/jit.h>.

No change in public API
