+++ src/mono/mono/eventpipe/test/ep-teardown-tests.c
@@ -1,6 +1,6 @@
 #include <eventpipe/ep.h>
 #include <eglib/test/test.h>
-#include <mono/mini/jit.h>
+#include <mono/jit/jit.h>
 
 #define TEST_FILE "./ep_test_create_file.txt"
 #define TEST_FILE_2 "./ep_test_create_file_2.txt"
