+++ src/mono/mono/unit-tests/test-mono-callspec.c
@@ -26,7 +26,7 @@
 #include <mono/metadata/class-internals.h>
 #include <mono/metadata/appdomain.h>
 #include <mono/metadata/debug-helpers.h>
-#include <mono/mini/jit.h>
+#include <mono/jit/jit.h>
 #include <mono/utils/mono-error-internals.h>
 
 #define TESTPROG "callspec.exe"
