+++ src/mono/mono/unit-tests/test-mono-embed.c
@@ -3,5 +3,5 @@
  */
 
 #define _TESTCASE_
-#include <mono/mini/jit.h>
+#include <mono/jit/jit.h>
 #include <embed/teste.c>
