+++ src/native/public/mono/jit/details/jit-functions.h
