+++ src/native/public/mono/jit/details/jit-types.h
